Fixed Delete-Issue-feature: - fixed DeleteIssueCommand, DeleteHandler for Issue, IColumnRepository, ColumnRepository and SqlStatements; - deleted ColumnWithIssuesAndUserResponseDto; - fixed storage procedure in 02-SprocsForCrud.sql; - added a new storage procedure sp_Issues_GetByColumnIdWithUsersJson to 03-SprocsForBusinessLogic.sql.

+++ Board.BusinessLogic/Features/ForIssue/Commands/DeleteIssueCommand.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace Board.BusinessLogic.Features.ForIssue.Commands;
 
-public record DeleteIssueCommand(int Id) : IRequest<IssuesByColumnIdResponseDto>;
+public record DeleteIssueCommand(long Id) : IRequest<IssuesByColumnIdResponseDto>;
